fix: add safeguards to select component options assignment
