+++ frontend/src/lib/components/Forms/ModelForm/AppliedControlPolicyForm.svelte
@@ -9,6 +9,7 @@
 	import type { SuperValidated } from 'sveltekit-superforms';
 	import type { ModelInfo, CacheLock } from '$lib/utils/types';
 	import { m } from '$paraglide/messages';
+	import { onMount } from 'svelte';
 
 	export let form: SuperValidated<any>;
 	export let model: ModelInfo;
@@ -18,11 +19,21 @@
 	export let schema: any = {};
 	export let initialData: Record<string, any> = {};
 
-	if (model.selectOptions && 'priority' in model.selectOptions) {
-		model.selectOptions['priority'].forEach((element) => {
-			element.value = parseInt(element.value);
-		});
-	}
+	onMount(async () => {
+		const selectOptions = {
+			status: await fetch('/applied-controls/status').then((r) => r.json()),
+			priority: await fetch('/applied-controls/priority').then((r) => r.json()),
+			category: await fetch('/applied-controls/category').then((r) => r.json()),
+			csf_function: await fetch('/applied-controls/csf_function').then((r) => r.json()),
+			effort: await fetch('/applied-controls/effort').then((r) => r.json())
+		};
+		model.selectOptions = selectOptions;
+		if (model.selectOptions && 'priority' in model.selectOptions) {
+			model.selectOptions['priority'].forEach((element) => {
+				element.value = parseInt(element.value);
+			});
+		}
+	});
 </script>
 
 {#if !duplicate}
@@ -38,7 +49,7 @@
 	/>
 	<Select
 		{form}
-		options={model.selectOptions['status']}
+		options={model.selectOptions?.status}
 		field="status"
 		label={m.status()}
 		cacheLock={cacheLocks['status']}
@@ -81,7 +92,7 @@
 		/>
 		<Select
 			{form}
-			options={model.selectOptions['priority']}
+			options={model.selectOptions?.priority}
 			field="priority"
 			label={m.priority()}
 			cacheLock={cacheLocks['priority']}
@@ -111,7 +122,7 @@
 		{#if schema.shape.category}
 			<Select
 				{form}
-				options={model.selectOptions['category']}
+				options={model.selectOptions?.category}
 				field="category"
 				label={m.category()}
 				cacheLock={cacheLocks['category']}
@@ -120,7 +131,7 @@
 		{/if}
 		<Select
 			{form}
-			options={model.selectOptions['csf_function']}
+			options={model.selectOptions?.csf_function}
 			field="csf_function"
 			label={m.csfFunction()}
 			cacheLock={cacheLocks['csf_function']}
@@ -146,7 +157,7 @@
 		/>
 		<Select
 			{form}
-			options={model.selectOptions['effort']}
+			options={model.selectOptions?.effort}
 			field="effort"
 			label={m.effort()}
 			helpText={m.effortHelpText()}
@@ -155,7 +166,7 @@
 		/>
 		<Select
 			{form}
-			options={model.selectOptions['control_impact']}
+			options={model.selectOptions?.control_impact}
 			field="control_impact"
 			label={m.controlImpact()}
 			helpText={m.impactHelpText()}
